ci: revert remove all github actions test workflows
All tests were removed in https://github.com/etcd-io/etcd/pull/18058,
https://github.com/etcd-io/etcd/pull/18249, https://github.com/etcd-io/etcd/pull/18454,
https://github.com/etcd-io/etcd/pull/18714, https://github.com/etcd-io/etcd/pull/19550
and https://github.com/etcd-io/etcd/pull/19551 as part of the upstream
migration to prow https://github.com/kubernetes/test-infra/issues/32754.

Since we maintain a fork without tests running in prow, we want to
maintain the tests in github actions. Hence this commit reverts the
three pull requests mentioned above with the exception of release.yaml
since we do not use the docker images built in this repository and a
bump to golangci-lint v2.

Signed-off-by: Ayaz Badouraly <ayaz.badouraly@datadoghq.com>